Added a formatter class
